% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snakemake.R
\name{snakemake}
\alias{snakemake}
\alias{run_snakemake}
\alias{create_dag}
\title{Snakemake wrapper}
\usage{
snakemake(rules, dryrun = F, forceall = F, code_changes = T,
  cores = "all", snakefile = "Snakefile", args = c("-p"),
  async = F)

run_snakemake(args)

create_dag(args = c("-p"), filename = "dag.svg")
}
\arguments{
\item{rules}{rules to run}

\item{cores}{number of cores to use}

\item{args}{any additional valid snakemake arguments are being passed}

\item{dryRun}{boolean flag indicating test run}

\item{args}{}
}
\description{
Snakemake wrapper
}
